refactor(perl-ci-hygiene): split version sync into srp submodules (#0000)
